requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
keras>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.12.0
